% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Median.R
\name{Median}
\alias{Median}
\title{This function is a wrapper enabling \code{na.rm=TRUE} for the stats::median function. This function may
be used in the \code{tabular} function of the \code{tables}-package.}
\usage{
Median(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
median of the input vector while disregarding NAs
}
\description{
This function is a wrapper enabling \code{na.rm=TRUE} for the stats::median function. This function may
be used in the \code{tabular} function of the \code{tables}-package.
}
\examples{
vec <- c(1, 2, 3, NA, NA, 4, 5)
Median(vec)

}
