% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_level.R
\name{na_to_level}
\alias{na_to_level}
\title{This functions replaces all NA's in a factor to a level with specified name.}
\usage{
na_to_level(x, name = "N/A")
}
\arguments{
\item{x}{a factor}

\item{name}{character - the name of the new level (default: N/A)}
}
\value{
The same factor with renamed NA
}
\description{
This functions replaces all NA's in a factor to a level with specified name.
}
\examples{
na_to_level(factor(c("a","b", "b", NA))) # Levels: a b N/A
}
