% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.R
\name{Sum}
\alias{Sum}
\title{This function is a wrapper enabling \code{na.rm=TRUE} for the sum function. This function may
be used in the \code{tabular} function of the \code{tables}-package.}
\usage{
Sum(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Return value of the Sum function.
}
\description{
This function is a wrapper enabling \code{na.rm=TRUE} for the sum function. This function may
be used in the \code{tabular} function of the \code{tables}-package.
}
